fix(gpu): ten kernel args the package build is structurally unable to see
`Int`/`UInt` stopped conforming to `DevicePassable` in 1.0.0rc2 (`Bool` never
did). A kernel taking one compiles happily; it fails where it is LAUNCHED,
as a constraint error thrown from inside `SIMD`:

    simd.mojo:612: constraint failed: Int and UInt do not conform to
    DevicePassable; use a fixed-width type such as Int32 or Int64 instead

`pixi run build` cannot catch this, and not by accident: `mojo precompile`
stops at elaboration and never instantiates a generic kernel, so the failure
only exists once some test or example picks the parameters. `mcts_gpu.mojo`
sat broken behind a green build until `test_az_connect4_arena` was run.

Ten parameters across eight kernels, each now `Int32`, re-widened with
`Int(...)` at its uses and cast at every launch site — a bare `Int` argument
re-introduces the failure through implicit conversion even when the signature
is right:

  _atari_reset_kernel / _atari_step_kernel   rom_size, n_envs
  _imag_ret_k                                slowtar
  _scale_kernel (gpt)                        n
  gather_rows_kernel                         offset
  gpu_mcts_extract_actions_temp_kernel       temp_threshold
  gz_extract_actions_temp_kernel             temp_threshold
  ilqr_apply_control_update_kernel           t

All ten are small — counts, flags, a horizon index — so `Int32` is the same
value, not a truncation. `mppi_sample_actions_batched_kernel` was fixed this
way at rc2 and is the shape followed here. Two of these had already met the
rule and mis-generalised from it: `_imag_ret_k` carried the comment "`slowtar`
is Int not Bool — Bool isn't a valid GPU kernel scalar arg", and
`extract_actions_temp`'s docstring said `TEMP_THRESHOLD` is comptime *because*
`enqueue_function` rejects runtime `Int` — then passed it as a runtime arg
anyway. Both are corrected.

The real deliverable is `scripts/audit_kernel_scalar_args.py`: the build-time
check the compiler cannot give us. It resolves every symbol reaching
`enqueue_function[...]` (686 of them, 686 resolved) and flags non-DevicePassable
scalars; exit code is the finding count, so it can gate CI. It reports
unresolved launches as findings rather than passing over them, because a
launch it cannot trace is a blind spot, not a pass — the first draft mapped
each `comptime k = ...` alias with a dict, so same-named aliases collapsed to
the last binding and it silently missed `_atari_step_kernel` and
`gz_extract_actions_temp_kernel`. It now maps an alias to the SET of its
bindings. Verified non-vacuous by reintroducing a defect and watching it fire.

Verified: az_connect4_arena compiles AND trains (93 -> 107 wins vs the same
opponent); ilqr GPU oracle + compile, az_gumbel arena + search, gpt smoke,
trainer augment/shuffle smoke, dreamerv3 cartpole GPU smoke — 7/7. Package
build 0 errors, 0 warnings. examples-compile 19/19.

Not fixed here, found while verifying and unrelated to this class:
`examples/arcade_games/atari_gpu_env_parity.mojo` and
`rainbow_atari_pong_ram_training_gpu.mojo` fail to parse on a different 1.0
tightening — `Pointer[UInt8, MutUntrackedOrigin]` no longer converts to
`MutAnyOrigin` at `AtariEnvironment.__init__`. Reproduced with the pre-`alloc`
`environment.mojo`, so it predates that commit. Neither file is in any
manifest, which is why nothing caught it.

+++ mojo_rl/deep_agents/training/atari_gpu_env.mojo
@@ -69,15 +69,18 @@ def _atari_reset_kernel[
     states: Pointer[AtariState, MutAnyOrigin],
     s0: Pointer[AtariState, MutAnyOrigin],
     rom: Pointer[UInt8, MutAnyOrigin],
-    rom_size: Int,
+    # ⚠ Int32, NOT Int — `Int`/`UInt` are not `DevicePassable`. A bare
+    # `Int` still compiles in `pixi run build` and fails only where the
+    # kernel is LAUNCHED; keep the `Int32(...)` casts at the call sites.
+    rom_size: Int32,
     op_table: Pointer[OpcodeEntry, MutAnyOrigin],
     dones: Pointer[Scalar[DT], MutAnyOrigin],
     obs: Pointer[Scalar[DT], MutAnyOrigin],
-    n_envs: Int,
+    n_envs: Int32,
     seed: UInt64,
 ):
     var i = Int(global_idx.x)
-    if i >= n_envs:
+    if i >= Int(n_envs):
         return
     comptime if SELECTIVE:
         if dones[unsafe_offset=i] <= 0.5:
@@ -89,7 +92,7 @@ def _atari_reset_kernel[
         for _ in range(k):
             set_action(st, ACTION_NOOP)
             run_frame_cycle_accurate[RENDER=False](
-                st, rom, rom_size, dummy.unsafe_ptr(), op_table
+                st, rom, Int(rom_size), dummy.unsafe_ptr(), op_table
             )
     # Rebase: episode reward baseline + frame budget start fresh post-decorrelation.
     st.score = Int32(GAME.get_score(st.ram))
@@ -109,17 +112,20 @@ def _atari_step_kernel[
 ](
     states: Pointer[AtariState, MutAnyOrigin],
     rom: Pointer[UInt8, MutAnyOrigin],
-    rom_size: Int,
+    # ⚠ Int32, NOT Int — `Int`/`UInt` are not `DevicePassable`. A bare
+    # `Int` still compiles in `pixi run build` and fails only where the
+    # kernel is LAUNCHED; keep the `Int32(...)` casts at the call sites.
+    rom_size: Int32,
     op_table: Pointer[OpcodeEntry, MutAnyOrigin],
     actions: Pointer[Scalar[DT], MutAnyOrigin],
     rewards: Pointer[Scalar[DT], MutAnyOrigin],
     dones: Pointer[Scalar[DT], MutAnyOrigin],
     terminated: Pointer[Scalar[DT], MutAnyOrigin],
     obs: Pointer[Scalar[DT], MutAnyOrigin],
-    n_envs: Int,
+    n_envs: Int32,
 ):
     var i = Int(global_idx.x)
-    if i >= n_envs:
+    if i >= Int(n_envs):
         return
     var st = states[unsafe_offset=i].copy()
     var ale = GAME.map_action(Int(actions[unsafe_offset=i]))
@@ -128,7 +134,7 @@ def _atari_step_kernel[
     for _ in range(FRAME_SKIP):
         set_action(st, ale)
         run_frame_cycle_accurate[RENDER=False](
-            st, rom, rom_size, dummy.unsafe_ptr(), op_table
+            st, rom, Int(rom_size), dummy.unsafe_ptr(), op_table
         )
     var new_score = GAME.get_score(st.ram)
     st.score = Int32(new_score)
@@ -272,11 +278,11 @@ struct AtariGpuBatchedEnv[
             self._states_p(),
             self._s0_p(),
             self._rom_p(),
-            self.rom_size,
+            Int32(self.rom_size),
             self._opt_p(),
             mptr(self._done.unsafe_ptr()),
             mptr(self._obs.unsafe_ptr()),
-            Self.N_ENVS,
+            Int32(Self.N_ENVS),
             rng_seed,
             grid_dim=(blocks,),
             block_dim=(_TPB,),
@@ -295,14 +301,14 @@ struct AtariGpuBatchedEnv[
         c.enqueue_function[k](
             self._states_p(),
             self._rom_p(),
-            self.rom_size,
+            Int32(self.rom_size),
             self._opt_p(),
             mptr(self._action.unsafe_ptr()),
             mptr(self._reward.unsafe_ptr()),
             mptr(self._done.unsafe_ptr()),
             mptr(self._terminated.unsafe_ptr()),
             mptr(self._obs.unsafe_ptr()),
-            Self.N_ENVS,
+            Int32(Self.N_ENVS),
             grid_dim=(blocks,),
             block_dim=(_TPB,),
         )
@@ -318,11 +324,11 @@ struct AtariGpuBatchedEnv[
             self._states_p(),
             self._s0_p(),
             self._rom_p(),
-            self.rom_size,
+            Int32(self.rom_size),
             self._opt_p(),
             mptr(self._done.unsafe_ptr()),
             mptr(self._obs.unsafe_ptr()),
-            Self.N_ENVS,
+            Int32(Self.N_ENVS),
             rng_seed,
             grid_dim=(blocks,),
             block_dim=(_TPB,),
