+++ mojo_rl/nn/models/gpt.mojo
@@ -228,11 +228,16 @@ comptime GPTDropTied[
 
 
 def _scale_kernel(
-    buf: Pointer[Scalar[DT], MutAnyOrigin], n: Int, s: Scalar[DT]
+    buf: Pointer[Scalar[DT], MutAnyOrigin],
+    # ⚠ Int32, NOT Int — `Int`/`UInt` are not `DevicePassable`. A bare
+    # `Int` still compiles in `pixi run build` and fails only where the
+    # kernel is LAUNCHED; keep the `Int32(...)` casts at the call sites.
+    n: Int32,
+    s: Scalar[DT],
 ):
     """`buf[i] *= s`. One thread per element."""
     var i = Int(global_idx.x)
-    if i < n:
+    if i < Int(n):
         buf[unsafe_offset=i] = buf[unsafe_offset=i] * s
 
 
@@ -256,7 +261,7 @@ def _gpt_scale_weight[target: StaticString, N: Int](
     else:
         var c = ctx.value()
         c.enqueue_function[_scale_kernel](
-            w.dev.value(), N, s,
+            w.dev.value(), Int32(N), s,
             grid_dim=(N + TPB - 1) // TPB, block_dim=TPB,
         )
